[fs] Make the timespec32 narrowing explicit (-Wshorten-64-to-32)
devfs.c:238 and fat32.c:341-342 assigned a time_t value into
k_timespec32::tv_sec (s32). The narrowing is deliberate --
struct k_stat64 documents its Y2038 limitation in sys_types.h
-- but the implicit conversion trips -Wshorten-64-to-32 under
clang_wconv's -Werror. Add explicit (s32) casts, mirroring the
fix already applied to modules/sysfs/sysfs.c in c9a2ef08.

Pre-existing issue; only surfaced now because the clang_wconv
generator now reaches these TUs on darwin hosts.

+++ kernel/fs/devfs.c
@@ -235,7 +235,14 @@ devfs_stat(struct mnt_fs *fs, vfs_inode_ptr_t i, struct k_stat64 *statbuf)
    statbuf->st_blksize = PAGE_SIZE;
    statbuf->st_blocks = 0;
 
-   statbuf->st_ctim.tv_sec = ddata->wrt_time;
+   /*
+    * Narrowing to k_timespec32::tv_sec is part of the Y2038
+    * limitation of struct k_stat64 -- see the comment on the
+    * struct declaration in include/tilck/kernel/sys_types.h.
+    * Make the cast explicit so -Wshorten-64-to-32 (clang_wconv)
+    * doesn't trip on what's a deliberate truncation.
+    */
+   statbuf->st_ctim.tv_sec = (s32)ddata->wrt_time;
    statbuf->st_mtim = statbuf->st_ctim;
    statbuf->st_atim = statbuf->st_mtim;
